Address review comment: split execute workflow into multiple files
